multimodal-museum-guide/requirements.txt
Flask
gTTS
googletrans==4.0.0-rc1
groq
python-dotenv